Change file.cc and directory.cc to use sqlite.

+++ fs/version.h
@@ -23,11 +23,10 @@
 #include "ndnfs.h"
 #include "segment.h"
 
-inline long long generate_version() {
-    //std::cout << "generate_version: called" << std::endl;
+inline uint64_t generate_version() {
     struct timeval tv;
     gettimeofday(&tv,NULL);
-    return (long long)(tv.tv_sec * (uint64_t)1000000 + tv.tv_usec);
+    return (uint64_t)(tv.tv_sec * (uint64_t)1000000 + tv.tv_usec);
 }
